Create mini shell
Program Description:
This program is a mini-shell. The user can open a terminal and type commands to be executed. The shell uses a new process to run the commands.

#ex2 update: Now this shell support one pipe ("|") and redirections (<, >, >>, 2>). 
Also support a combination of pipe and redirections. 
The commands will be write into log.txt file, so note to attache it while run the shell (see instruction below).

Compilation:
Type in the terminal: gcc ex2.c -o ex2
#ex2 update: Note! Before compiling the file, make sure to delete makefile, so cmd: "make" will print an error.

Launch the Program: 
After compilation to ex2, in the terminal type: ./ex2 log.txt

NOTE:
This shell CAN NOT run the "cd" command. 
Also I fixed some bug with spaces-only string and some crash made by fgets() and other functions.